Move compat function package
